move event tabs in table and connect to api
