+++ src/DetailView/DetailView.js
@@ -1,23 +1,13 @@
 import React, { Component } from 'react';
-import {connect, PromiseState} from 'react-refetch';
-import ConnectionComponent from './../Utils/ConnectionComponent.js';
 import EntityInformation from './EntityInformation';
 import EventDiagram from './EventDiagram';
 import SearchBar from './../Utils/SearchBar';
-import EventTabs from './EventTabs';
 import EventTable from './EventTable';
 import { css } from 'aphrodite';
 import AppStyles from './../AppStyles';
 
-class DetailView extends ConnectionComponent {
+class DetailView extends Component {
 	render() {
-        const allFetches = PromiseState.all([this.props.entity]);
-        const entity = this.props.entity.value;
-        const eventTypes = this.props.eventTypes.value;
-        const connectionIncomplete = super.render(allFetches);
-        if(connectionIncomplete) {
-            return connectionIncomplete;
-        }
         return (
 			<div>
 				<div className={css(AppStyles.dFlex, AppStyles.elementMarginTop)}>
@@ -25,15 +15,11 @@ class DetailView extends ConnectionComponent {
 					<EventDiagram styles={[AppStyles.w50]}/>
 				</div>
 				<SearchBar styles={[AppStyles.elementMarginTop]}/>
-				<EventTabs styles={[AppStyles.elementMarginTop]}/>
-				<EventTable/>
+				<EventTable entityId={this.props.match.params.entityId} />
 			</div>
 		);
 	}
 }
 
 
-export default connect.defaults({fetch: ConnectionComponent.switchFetch})(props => ({
-    entity: `/entity/${props.match.params.entityId}`,
-	eventTypes: `/entity/${props.match.params.entityId}/eventtypes`,
-}))(DetailView);
+export default DetailView;
